feat: Qr generator repositoy
