+++ androidapp/app/src/main/java/com/juandgaines/seedqrvalidator/home/presentation/HomeViewModel.kt
@@ -2,15 +2,12 @@ package com.juandgaines.seedqrvalidator.home.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.juandgaines.seedqrvalidator.core.domain.HomeRepository
-import com.juandgaines.seedqrvalidator.core.domain.utils.onSuccess
+import com.juandgaines.seedqrvalidator.home.domain.HomeRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
