+++ androidapp/app/src/main/java/com/juandgaines/seedqrvalidator/reader/presentation/QRGeneratorScreen.kt
@@ -3,6 +3,8 @@ package com.juandgaines.seedqrvalidator.reader.presentation
 import androidx.compose.runtime.Composable
 
 @Composable
-fun ReaderScreenRoot(){
+fun QrGeneratorScreenRoot(
+
+){
 
 }
