+++ androidapp/app/src/main/java/com/juandgaines/seedqrvalidator/reader/presentation/QRGeneratorViewModel.kt
@@ -1,0 +1,9 @@
+package com.juandgaines.seedqrvalidator.reader.presentation
+
+import com.juandgaines.seedqrvalidator.home.domain.HomeRepository
+import javax.inject.Inject
+
+class QRGeneratorViewModel @Inject constructor(
+    private val seedRepository: HomeRepository
+){
+}
